requirements.txt
absl-py==0.9.0
aiohttp @ file:///tmp/build/80754af9/aiohttp_1603893234148/work
appdirs==1.4.3
argon2-cffi @ file:///tmp/build/80754af9/argon2-cffi_1596828452693/work
astor==0.8.1
async-generator==1.10
async-timeout==3.0.1
atari-py==0.2.6
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
autopep8==1.5.5
backcall==0.2.0
bcolz==1.2.1
beautifulsoup4==4.9.3
bleach @ file:///tmp/build/80754af9/bleach_1600439572647/work
blinker==1.4
blis==0.4.1
bokeh==2.3.1
boto==2.49.0
boto3==1.9.66
botocore==1.12.67
Box2D==2.3.10
cachetools @ file:///tmp/build/80754af9/cachetools_1596822027882/work
catalogue==1.0.0
certifi==2021.5.30
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1602537219008/work
chardet @ file:///tmp/build/80754af9/chardet_1605303159953/work
ckiptagger==0.1.1
clearbit==0.1.7
click==7.1.2
cloudpickle @ file:///home/conda/feedstock_root/build_artifacts/cloudpickle_1598400192773/work
cryptography @ file:///tmp/build/80754af9/cryptography_1605544449973/work
cycler==0.10.0
cymem==2.0.3
Cython==0.29.17
cytoolz==0.11.0
dask @ file:///home/conda/feedstock_root/build_artifacts/dask-core_1602029610262/work
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
docutils==0.16
easydict==1.9
en-core-web-sm==2.2.5
entrypoints==0.3
filelock==3.0.12
fsspec @ file:///tmp/build/80754af9/fsspec_1602684995936/work
future==0.18.2
gast==0.2.2
gdown==3.10.3
gensim==3.8.0
glfw==1.11.0
google-api-core @ file:///tmp/build/80754af9/google-api-core-split_1601307042068/work
google-auth @ file:///tmp/build/80754af9/google-auth_1604100403705/work
google-auth-oauthlib @ file:///tmp/build/80754af9/google-auth-oauthlib_1603929124518/work
google-cloud-core @ file:///tmp/build/80754af9/google-cloud-core_1605284563418/work
google-cloud-storage @ file:///tmp/build/80754af9/google-cloud-storage_1603300774238/work
google-crc32c @ file:///tmp/build/80754af9/google-crc32c_1601305107307/work
google-pasta==0.2.0
google-resumable-media @ file:///tmp/build/80754af9/google-resumable-media_1601995638063/work
googleapis-common-protos @ file:///tmp/build/80754af9/googleapis-common-protos-feedstock_1600455614392/work
grpcio==1.28.1
gym==0.17.1
h5py==2.10.0
helper==2.4.2
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
imagecodecs-lite @ file:///home/conda/feedstock_root/build_artifacts/imagecodecs-lite_1602583028496/work
imageio @ file:///home/conda/feedstock_root/build_artifacts/imageio_1594044661732/work
importlib-metadata==1.6.0
imutils==0.5.4
ipykernel==5.1.4
ipython @ file:///tmp/build/80754af9/ipython_1604101195213/work
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi @ file:///tmp/build/80754af9/jedi_1596490743326/work
jieba==0.39
Jinja2==2.11.2
jmespath @ file:///tmp/build/80754af9/jmespath_1594304593830/work
joblib @ file:///tmp/build/80754af9/joblib_1601912903842/work
json5==0.9.4
jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1601311786391/work
jupyter-core==4.6.3
jupyterlab==2.1.0
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab-server==1.1.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver @ file:///tmp/build/80754af9/kiwisolver_1604014532738/work
livelossplot==0.5.4
lockfile==0.12.2
Markdown==3.2.1
MarkupSafe @ file:///tmp/build/80754af9/markupsafe_1594371495811/work
matplotlib==3.1.3
mistune @ file:///tmp/build/80754af9/mistune_1594373098390/work
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
multidict @ file:///tmp/build/80754af9/multidict_1600456400975/work
murmurhash==1.0.2
nbclient @ file:///tmp/build/80754af9/nbclient_1602783176460/work
nbconvert @ file:///tmp/build/80754af9/nbconvert_1601914821128/work
nbformat @ file:///tmp/build/80754af9/nbformat_1602783287752/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1605115881283/work
networkx @ file:///home/conda/feedstock_root/build_artifacts/networkx_1598210780226/work
nltk==3.4.5
notebook @ file:///tmp/build/80754af9/notebook_1601501580008/work
numpy==1.19.0
oauthlib==3.1.0
olefile==0.46
opencv-contrib-python==4.5.1.48
opencv-python==4.2.0.34
opt-einsum==3.2.1
packaging==20.4
pandas @ file:///tmp/build/80754af9/pandas_1602088128026/work
pandocfilters @ file:///tmp/build/80754af9/pandocfilters_1605120451932/work
parso==0.7.0
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickle-mixin==1.0.2
pickleshare @ file:///tmp/build/80754af9/pickleshare_1594384075987/work
Pillow==7.1.0
plac==1.1.3
preshed==3.0.2
prometheus-client==0.8.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
protobuf==3.11.3
psutil==5.7.2
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1605560620615/work/dist/ptyprocess-0.6.0-py2.py3-none-any.whl
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyglet==1.5.0
Pygments @ file:///tmp/build/80754af9/pygments_1604103097372/work
PyJWT==1.7.1
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1605545627475/work
pyparsing==2.4.7
pyrsistent @ file:///tmp/build/80754af9/pyrsistent_1600141707582/work
PySocks==1.7.1
pytesseract==0.3.7
python-dateutil==2.8.1
pytorch-lightning @ file:///home/conda/feedstock_root/build_artifacts/pytorch-lightning_1602786328955/work
pytz==2020.1
PyWavelets @ file:///home/conda/feedstock_root/build_artifacts/pywavelets_1602504439149/work
PyYAML==5.3.1
pyzmq==19.0.2
regex==2020.2.20
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
requests-oauthlib==1.3.0
rouge-score==0.0.3
rsa @ file:///tmp/build/80754af9/rsa_1596998415516/work
s3transfer==0.1.13
sacremoses==0.0.43
scikit-image==0.17.2
scikit-learn @ file:///tmp/build/80754af9/scikit-learn_1598376882706/work
scipy @ file:///tmp/build/80754af9/scipy_1597686620742/work
seaborn @ file:///tmp/build/80754af9/seaborn_1600553570093/work
selenium==3.141.0
Send2Trash==1.5.0
sentence-transformers==0.4.1.2
sentencepiece==0.1.91
six==1.14.0
skorch @ file:///home/conda/feedstock_root/build_artifacts/skorch_1591873157816/work
smart-open==2.0.0
soupsieve==2.1
spacy==2.2.4
srsly==1.0.2
stanfordcorenlp==3.9.1.1
tabulate==0.8.7
tensorboard==1.15.0
tensorboard-plugin-wit==1.6.0
tensorboardX==2.1
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
thinc==7.4.0
threadpoolctl @ file:///tmp/tmp9twdgx9k/threadpoolctl-2.1.0-py3-none-any.whl
tifffile==2019.7.26.2
tokenizers==0.10.1
toml==0.10.2
toolz @ file:///home/conda/feedstock_root/build_artifacts/toolz_1600973991856/work
torch==1.7.1
torchvision==0.5.0+cu100
tornado==6.0.4
tqdm==4.45.0
traitlets==4.3.3
transformers==4.3.3
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1598376058250/work
urllib3==1.25.9
virtualenv==20.0.15
wasabi==0.6.0
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webcolors==1.11.1
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
yarl @ file:///tmp/build/80754af9/yarl_1598045053219/work
zipp==3.1.0
